# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API сервиса iAmOkey",
        "description": "Документация сервиса.",
        "contact": {
            "email": "nfs2025@mail.ru"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/indicator/{patient_id}": {
            "get": {
                "tags": [
                    "Indicator"
                ],
                "summary": "Показатели пациента",
                "description": "Возвращается объект всех показаний по пациенту",
                "operationId": "0b837330570cca3e5afea65bb9b8a539",
                "parameters": [
                    {
                        "name": "patient_id",
                        "in": "path",
                        "description": "ID пациента",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "description": "Дата начала",
                        "required": true
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "description": "Дата конца",
                        "required": true
                    },
                    {
                        "name": "period",
                        "in": "query",
                        "description": "day|week|month|year",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/indicator/": {
            "put": {
                "tags": [
                    "Indicator"
                ],
                "summary": "Создать нового запись данных показаний",
                "description": "Создать нового запись данных показаний, возвращается объект показаний",
                "operationId": "0964db250ca24083042397fefaebc30d",
                "requestBody": {
                    "$ref": "#/components/requestBodies/VehicleStoreRequestBody"
                },
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/medical/{patient_id}": {
            "get": {
                "tags": [
                    "Medical"
                ],
                "summary": "Медикаменты пользователя",
                "description": "Возвращается объект медикаментов",
                "operationId": "02e1cb72f102b46dbb70819b2ae5745a",
                "parameters": [
                    {
                        "name": "patient_id",
                        "in": "path",
                        "description": "patient_id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/medical/": {
            "put": {
                "tags": [
                    "Medical"
                ],
                "summary": "Создать новый медикамент",
                "description": "Создатьновый медикамент, возвращается объект медикамент",
                "operationId": "faa473606f0fe8e6e02ae47c6d3262c6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Medical"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Medical"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/medical/{id}": {
            "delete": {
                "tags": [
                    "Medical"
                ],
                "summary": "Удалить медикамент",
                "description": "Удалить медикамент",
                "operationId": "262fbc0a712442e7a9a3c72b873594b7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id медикамента",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/medicallog/{patient_id}": {
            "get": {
                "tags": [
                    "MedicalLog"
                ],
                "summary": "Прием таблеток пациента",
                "description": "Возвращается объект приема таблеток пациента",
                "operationId": "8714ea447f60b7388919eb4cf56ed663",
                "parameters": [
                    {
                        "name": "patient_id",
                        "in": "path",
                        "description": "patient_id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/medicallog/": {
            "put": {
                "tags": [
                    "MedicalLog"
                ],
                "summary": "Запись приема таблеток",
                "description": "Создать нового пациента, возвращается объект приема таблеток пациента",
                "operationId": "02be5474c360468c111b87ca4f484212",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MedicalLog"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MedicalLog"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/medicallog/{id}": {
            "delete": {
                "tags": [
                    "MedicalLog"
                ],
                "summary": "Удалить запись приема таблеток",
                "description": "Удалить запись приема таблеток",
                "operationId": "226929edca6e30e53ce4057143bc0b98",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id уведомления",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Medical": {
                "title": "Medical",
                "description": "Medical model",
                "properties": {
                    "patient_id": {
                        "title": "Patient_id",
                        "description": "Id пациента",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "Name",
                        "description": "Название медикамента",
                        "type": "integer",
                        "example": "Глицин"
                    },
                    "time": {
                        "title": "Time",
                        "description": "Время приема",
                        "type": "integer",
                        "example": "Глицин"
                    },
                    "type": {
                        "title": "Type",
                        "description": "Иконка (1,2,3)",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Medical"
                }
            },
            "MedicalLog": {
                "title": "MedicalLog",
                "description": "MedicalLog model",
                "properties": {
                    "medical_id": {
                        "title": "Medical_id",
                        "description": "Id пациента",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "status": {
                        "title": "Status",
                        "description": "0 - неуспешно, 1 - успешно",
                        "type": "string",
                        "example": 1
                    }
                },
                "type": "object",
                "xml": {
                    "name": "MedicalLog"
                }
            },
...
        },
...
        "examples": {
            "Indicator": {
                "summary": "Indicator",
                "value": [
                    {
                        "patient_id": 1,
                        "pressure": {
                            "systolic": 120,
                            "diastolic": 80
                        },
                        "pulse": {
                            "average": 80,
                            "min": 60,
                            "max": 90,
                            "dispersion": 96
                        },
                        "accelerometer": {
                            "x": 80,
                            "y": 60,
                            "z": 90
                        },
                        "saturation": 96,
                        "sleep": 5,
                        "steps": 60,
                        "time": "2021-12-29 09:23:40"
                    },
                    {
                        "patient_id": 1,
                        "pressure": {
                            "systolic": 120,
                            "diastolic": 80
                        },
                        "pulse": {
                            "average": 80,
                            "min": 60,
                            "max": 90,
                            "dispersion": 96
                        },
                        "accelerometer": {
                            "x": 80,
                            "y": 60,
                            "z": 90
                        },
                        "saturation": 96,
                        "sleep": 5,
                        "steps": 60,
                        "time": "2021-12-29 09:23:40"
                    }
                ]
            },
...
        },
        "requestBodies": {
            "VehicleStoreRequestBody": {
                "description": "Indicator",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {},
                        "examples": {
                            "#/components/examples/Indicator": {
                                "$ref": "#/components/examples/Indicator"
                            }
                        }
                    }
                }
            },
...
        }
    },
...
    "security": [
        []
    ]
}